Create "test.compile.and.run.only" and "@dot.init" targets.
Refactor @dot target by pulling out the declaration of @dot.classpath into a @dot.init target.  This will allow for its reuse later when running of the @dot target is not desired, but the @dot.classpath is needed. (When the tests are to run without the build).
Also, create an empty test.compile.and.run.only target so it can be called by the SeleniumPlus build.